feature flags: enable EnableAlterDatabaseCreateHiveFirst (#14078)
This is a long awaited improvement in the dedicated database creation process.

`EnableAlterDatabaseCreateHiveFirst` (`enable_alter_database_create_hive_first` in the config) feature flag affects how system tablets of the dedicated database are created (and than managed throughout their lifetime). System tablets (schemeshard, hive, coordinators/mediators, sysview-processor etc) are those that constitute and service the database itself.

Without `EnableAlterDatabaseCreateHiveFirst`, creation of the new dedicated database leaves all it's system tablets in the custody of the main ("root") Hive of the cluster (all other tablets created afterwards live in the custody of the tenant Hive). Database with that separation of responsibilities works fine. The downside is that complete lack of common vision and coordination between two Hives (root and the tenant's) can lead to suboptimal results when database' tablets (both system and others) are balanced under the load. 

With `EnableAlterDatabaseCreateHiveFirst`, creation of the new dedicated database leaves all tenant tablets (except tenant Hive) in the custody of the tenant hive. Which alleviates problem with tablet balancing.